Workaround for initialization order.
Made NewlineInput responsible for handling outputMirror, but that variable is not initialized before it's
called by the input setter.

As a not-great fix, wrap most of NewlineInput in a process.nextTick(). This is more than I want implementers
of `input` to do, but it simplifies figuring out how else to make bindings all work for removeListener.